Move myia.infer into myia.abstract
