+++ myia/abstract/graph_infer.py
