+++ myia/abstract/utils.py
