+++ myia/pipeline/resources.py
@@ -9,7 +9,7 @@ from .. import dtype, operations, parser, composite as C
 from ..specialize import TypeSpecializer
 from ..abstract.base import AbstractFunction
 from ..abstract.inf import from_value
-from ..infer import InferenceEngine, ANYTHING
+from ..abstract import InferenceEngine, ANYTHING
 from ..ir import Graph, clone
 from ..prim import ops as P
 from ..utils import overload, TypeMap, UNKNOWN
