+++ myia/pipeline/standard.py
@@ -3,7 +3,7 @@
 
 from ..compile import step_wrap_primitives, step_compile, step_link, \
     step_export
-from ..infer import Context
+from ..abstract import Context
 from ..ir import GraphManager
 from ..prim import py_implementations
 from ..abstract import AbstractTrack, abstract_inferrer_constructors
