+++ myia/validate.py
@@ -3,7 +3,7 @@
 from .dtype import Array, Tuple, List, Function, Number, Bool, Problem, \
     TypeMeta, TypeType, Class, External, EnvType, SymbolicKeyType, \
     JTagged, type_cloner, ismyiatype
-from .infer import DEAD
+from .abstract import DEAD
 from .ir import manage
 from .prim import Primitive, ops as P
 from .utils import overload, ErrorPool
