+++ tests/opt/test_opt.py
@@ -3,7 +3,7 @@ import pytest
 
 from myia import operations
 from myia.pipeline import scalar_pipeline
-from myia.infer import InferenceError
+from myia.abstract import InferenceError
 from myia.ir import Constant, isomorphic, GraphCloner
 from myia.opt import PatternSubstitutionOptimization as psub, \
     LocalPassOptimizer, pattern_replacer, sexp_to_graph, \
